requirements.txt
aiogram==2.25.2
asyncpg==0.30.0
python-dotenv==1.0.1
